refactor: validation & routes consolidation
