+++ routes/movies.js
@@ -1,7 +1,6 @@
 const router = require('express').Router();
-const { celebrate, Joi, Segments } = require('celebrate');
+const { deleteMovieValidation, updateMovieValidator } = require('../utils/validation');
 const { getMovies, deleteMovie, updateMovie } = require('../controllers/movies');
-const { validateEmail } = require('../utils');
 
 const ROOT = '/movies';
 
@@ -9,29 +8,11 @@ module.exports = router
   .get(ROOT, getMovies)
   .delete(
     `${ROOT}/:movieId`,
-    celebrate({
-      [Segments.PARAMS]: {
-        movieId: Joi.string().hex().length(24).required(),
-      },
-    }),
+    deleteMovieValidation,
     deleteMovie,
   )
   .post(
     ROOT,
-    celebrate({
-      [Segments.BODY]: Joi.object().keys({
-        country: Joi.string().required(),
-        director: Joi.string().required(),
-        duration: Joi.number().required(),
-        year: Joi.string().required(),
-        description: Joi.string().required(),
-        image: Joi.string().required().custom((value) => validateEmail(value)),
-        trailer: Joi.string().required().custom((value) => validateEmail(value)),
-        nameRU: Joi.string().required(),
-        nameEN: Joi.string().required(),
-        thumbnail: Joi.string().required().custom((value) => validateEmail(value)),
-        movieId: Joi.number().required(),
-      }),
-    }),
+    updateMovieValidator,
     updateMovie,
   );
