+++ routes/users.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { celebrate, Joi, Segments } = require('celebrate');
+const { updateProfileValidator } = require('../utils/validation');
 const { updateProfile, getUserInfo } = require('../controllers/users');
 
 const ROOT = '/users';
@@ -8,11 +8,6 @@ module.exports = router
   .get(`${ROOT}/me`, getUserInfo)
   .patch(
     `${ROOT}/me`,
-    celebrate({
-      [Segments.BODY]: Joi.object().keys({
-        email: Joi.string().email().min(2).required(),
-        name: Joi.string().min(2).max(30).required(),
-      }),
-    }),
+    updateProfileValidator,
     updateProfile,
   );
