--- routes/utils.js
@@ -1,5 +1,0 @@
-const router = require('express').Router();
-const { crashTest } = require('../controllers/utils');
-
-module.exports = router
-  .get('/crash-test', crashTest);
